chore: add SHA1 hash validation of the sysroot

+++ src/sysroot-fetcher.ts
@@ -1,3 +1,4 @@
+import * as crypto from 'crypto';
 import * as debug from 'debug';
 import * as fs from 'fs-extra';
 import * as path from 'path';
@@ -28,12 +29,20 @@ export async function downloadLinuxSysroot(electronVersion: string, targetArch: 
   const { Sha1Sum: sha, Tarball: fileName } = electronSysroots[`sid_${linuxArch}`];
   const sysrootURL = `${SYSROOT_BASE_URL}/${sha}/${fileName}`;
   let sysrootBuffer = await fetch(sysrootURL, 'buffer');
+
+  const actualSha = crypto.createHash('SHA1').update(sysrootBuffer).digest('hex');
+  d('expected sha:', sha);
+  d('actual sha:', actualSha);
+  if (sha !== actualSha) throw new Error(`Attempted to download the linux sysroot for ${electronVersion} but the SHA checksum did not match`);
+
+  d('decompressing sysroot');
   sysrootBuffer = await new Promise<Buffer>(resolve => lzma.decompress(sysrootBuffer, undefined, result => resolve(result)));
 
   const tmpTarFile = path.resolve(ELECTRON_GYP_DIR, `${electronVersion}-${fileName}`);
   if (await fs.pathExists(tmpTarFile)) await fs.remove(tmpTarFile);
   await fs.writeFile(tmpTarFile, sysrootBuffer);
 
+  d('extracting sysroot');
   await tar.x({
     file: tmpTarFile,
     cwd: sysrootDir,
